Dynamically setting timeline year width, allow timeline to move up and down. (#517)
* Fitting timeline decade layout into the size of the touch screen set in the app delegate configuration.

* Added timeline constraint, refactoring.

* Posting timeline vertical position.

* Updates for xcode 10.

* Fixed grouping bugs with timeline controls.

* Fixed multiple connection related bugs and improvements.

+++ MapExplorer-macOS/Sources/Controllers/TimelineViewController/Views/TimelineHeaderView.swift
@@ -32,7 +32,7 @@ class TimelineHeaderView: NSView {
     override func draw(_ dirtyRect: NSRect) {
         super.draw(dirtyRect)
         style.timelineBorderColor.setFill()
-        let tick = CGRect(x: CGFloat(style.timelineHeaderOffset), y: frame.height - Constants.tickHeight, width: Constants.tickWidth, height: Constants.tickHeight)
+        let tick = CGRect(x: style.timelineHeaderOffset, y: frame.height - Constants.tickHeight, width: Constants.tickWidth, height: Constants.tickHeight)
         let path = NSBezierPath(rect: tick)
         path.fill()
     }
